Updated network message handlers to use dependency injection. - Moved static message handler classes to individual classes with the new IMessageHandler interface. While it is possible for a message handler class to implement multiple IMessageHandler interfaces, currently the standard is to have a one to one relationship. - Updated any static singleton usages in network handlers to interfaces injected in constructor. - MessageHandlerAttribute is not longer required for message handlers, opcode is pulled from message attribute.

+++ Source/NexusForever.Game/Account/Unlock/GenericUnlockManager.cs
@@ -63,7 +63,7 @@ namespace NexusForever.Game.Account.Unlock
         public void UnlockAll(GenericUnlockType type)
         {
             foreach (GenericUnlockEntryEntry entry in GameTableManager.Instance.GenericUnlockEntry.Entries
-                .Where(e => e.GenericUnlockTypeEnum == (uint)type))
+                .Where(e => e.GenericUnlockTypeEnum == type))
             {
                 if (unlocks.ContainsKey(entry.Id))
                     continue;
